requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
chromadb==0.4.22
pypdf==4.0.0